training/requirements.txt
sagemaker>=2.200.0
boto3>=1.34.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0


